[SW-000] Rename de fonction

+++ app/src/main/java/com/example/starwars/model/Character.kt
@@ -19,6 +19,6 @@ data class Character(@PrimaryKey @JsonProperty("uid") var uid: Int,
                      @JsonProperty("edited") var edited: String) {
 
     init {
-        uid = url.retrieveIdFromURL().toInt()
+        uid = url.retrieveIdFromURL()
     }
 }
